requirements.txt
annotated-types==0.7.0
APScheduler==3.10.4
loguru==0.7.2
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.1
pydantic==2.10.3
pydantic-settings==2.6.1
pydantic_core==2.27.1
pytz==2024.2
rsa==4.9
six==1.17.0
Telethon==1.34.0
typing_extensions==4.12.2
tzlocal==5.2
